use signal based detectors
