--- src/bot/detectors/automation/index.ts
@@ -1,8 +1,0 @@
-import BrowserAutomationDetector from './browser-automation-detector';
-import ChromiumAutomationDetector from './chromium-automation-detector';
-
-export default function createAutomationDetectors(
-  env: detection.Environment
-): detection.Detector[] {
-  return [new BrowserAutomationDetector(env), new ChromiumAutomationDetector(env)];
-}
