+++ src/bot/detectors/environment/index.ts
@@ -1,0 +1,8 @@
+import HeadlessDetector from './headless-detector';
+import WebdriverDetector from './webdriver-detector';
+
+export default function createEnvironmentDetectors(
+  env: detection.Environment
+): detection.Detector[] {
+  return [new HeadlessDetector(env), new WebdriverDetector(env)];
+}
